[ci] Fix Tizen packaging + CodeQL integer-overflow on conv path
Two CI failures remaining on the q4_0-fp16 conv stack:

1. Tizen GBS rpmbuild: "Installed (but unpackaged) file(s) found:
   /usr/include/nntrainer/conv_indirect.h". The indirect-conv header is
   installed (it is transitively included by the public cpu_backend.h /
   compute_ops.h), so it must be packaged. Add it to %files devel next to
   cpu_backend.h/compute_ops.h.

2. CodeQL high: cpp/integer-multiplication-cast-to-long at
   conv2d_layer.cpp:781 -- 'icg * fh * fw' is computed in unsigned int and
   then widened to size_t for TensorDim's initializer_list<size_t>. Widen
   the first operand (matches the existing (size_t)g * ... offsets right
   below). Value-identical for all realistic dims; only removes the
   theoretical 32-bit overflow before the conversion.

+++ nntrainer/layers/conv2d_layer.cpp
@@ -780,7 +780,7 @@ void Conv2DLayer::forwarding(RunLayerContext &context, bool training) {
           ing_dim.setDataType(in_dt);
           Tensor in_g =
             in_sub.getSharedDataTensor(ing_dim, (size_t)g * icg * ihw);
-          TensorDim filtg_dim({ocg, icg * fh * fw});
+          TensorDim filtg_dim({ocg, (size_t)icg * fh * fw});
           filtg_dim.setDataType(filt_dt);
           Tensor filt_g = filter_kernel.getSharedDataTensor(
             filtg_dim, (size_t)g * ocg * icg * fh * fw);
